requirements.txt
python-dotenv
supabase
requests
brotli
pandas
lxml